requirements.txt
fastapi
uvicorn
uvicorn[standard]
pydantic[email]
email-validator
python-multipart
pymongo
pydantic_settings
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
httpx
pytest
